// src/conductor/providers/spotify/types/typed.ts
export declare namespace SpotifyApi {
// ...
  interface CurrentUsersProfileResponse extends UserObjectPrivate {}

  /**
   * Get User’s Followed Artists
   *
   * GET /v1/me/following
   * https://developer.spotify.com/web-api/get-followed-artists/
   */
  interface UsersFollowedArtistsResponse {
    artists: CursorBasedPagingObject<ArtistObjectFull>;
  }
// ...
  interface UserFollowsUsersOrArtistsResponse extends Array<boolean> {}
// ...
  interface UsersTopArtistsResponse extends PagingObject<ArtistObjectFull> {}

  /**
   * Get a User’s Top Artists and Tracks (Note: This is only Tracks)
   *
   * GET /v1/me/top/{type}
   * https://developer.spotify.com/web-api/get-users-top-artists-and-tracks/
   */
  interface UsersTopTracksResponse extends PagingObject<TrackObjectFull> {}
// ...
  interface UserProfileResponse extends UserObjectPublic {}
// ...
  interface UsersFollowPlaylistResponse extends Array<boolean> {}
// ...
  interface ArtistObjectFull extends ArtistObjectSimplified {
    /**
     * Information about the followers of the artist.
     */
    followers: FollowersObject;
    /**
     * A list of the genres the artist is associated with.
     * For example: `"Prog Rock"` , `"Post-Grunge"`.
     * (If not yet classified, the array is empty.)
     */
    genres: string[];
    /**
     * Images of the artist in various sizes, widest first.
     */
    images: ImageObject[];
    /**
     * The popularity of the artist. The value will be between `0` and `100`, with `100` being the most popular.
     * The artist’s popularity is calculated from the popularity of all the artist’s tracks.
     */
    popularity: number;
  }
// ...
  interface PagingObject<T> {
    href: string;
    items: T[];
    limit: number;
    next: string | null;
    offset: number;
    previous: string | null;
    total: number;
  }

  /**
   * Cursor Based Paging Object wrappers used for retrieving collections from the Spotify API.
   * [](https://developer.spotify.com/web-api/object-model/#cursor-based-paging-object)
   */
  interface CursorBasedPagingObject<T> {
    href: string;
    items: T[];
    limit: number;
    next: string | null;
    cursors: CursorObject;
    total?: number | undefined;
  }
// ...
  interface TrackObjectFull extends TrackObjectSimplified {
    /**
     * The album on which the track appears.
     */
    album: AlbumObjectSimplified;
    /**
     * Known external IDs for the track.
     */
    external_ids: ExternalIdObject;
    /**
     * The popularity of the track. The value will be between `0` and `100`, with `100` being the most popular.
     * The popularity of a track is a value between `0` and `100`, with `100` being the most popular.
     * The popularity is calculated by algorithm and is based, in the most part,
     * on the total number of plays the track has had and how recent those plays are.
     */
    popularity: number;
    /**
     * Whether or not the track is from a local file.
     */
    is_local?: boolean | undefined;
  }
// ...
  interface UserObjectPrivate extends UserObjectPublic {
    birthdate: string;
    country: string;
    email: string;
    product: string;
  }

  /**
   * User Object (Public)
   * [](https://developer.spotify.com/web-api/object-model/#track-object-simplified)
   */
  interface UserObjectPublic {
    display_name?: string | undefined;
    external_urls: ExternalUrlObject;
    followers?: FollowersObject | undefined;
    href: string;
    id: string;
    images?: ImageObject[] | undefined;
    type: 'user';
    uri: string;
  }
// ...
}